+++ nx/lib/nx/defn/expr.ex
@@ -1184,18 +1184,6 @@ defmodule Nx.Defn.Expr do
   end
 
   @impl true
-  def take(out, tensor, indices, axis) do
-    {[tensor, indices], context} = to_exprs([tensor, indices])
-    expr(out, context, :take, [tensor, indices, axis])
-  end
-
-  @impl true
-  def take_along_axis(out, tensor, indices, axis) do
-    {[tensor, indices], context} = to_exprs([tensor, indices])
-    expr(out, context, :take_along_axis, [tensor, indices, axis])
-  end
-
-  @impl true
   def gather(out, tensor, indices, opts) do
     {[tensor, indices], context} = to_exprs([tensor, indices])
     expr(out, context, :gather, [tensor, indices, opts])
@@ -1211,6 +1199,12 @@ defmodule Nx.Defn.Expr do
   def concatenate(out, tensors, axis) do
     {tensors, context} = to_exprs(tensors)
     expr(out, context, :concatenate, [tensors, axis])
+  end
+
+  @impl true
+  def stack(out, tensors, axis) do
+    {tensors, context} = to_exprs(tensors)
+    expr(out, context, :stack, [tensors, axis])
   end
 
   @impl true